+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## 0.0.1 (UNRELEASED)
   - Empty Bootstrap of Logstash filter plugin [#1](https://github.com/logstash-plugins/logstash-filter-elastic_integration/pull/1)
+  - Adds basic configuration options required for Elasticsearch connection [#2](https://github.com/logstash-plugins/logstash-filter-elastic_integration/pull/2)